==> notebooks/00_exploration.ipynb <==
# <center>  **Data Exploration**

------

# Make Constant Variables For Data Directory to access Data Faster in Notebook
DATA_PATH_CHILD  = '../data/raw/G2000_data.csv'
DATA_PATH_PARENT = '../data/raw/EthA_data.csv'

##  Import libraries

import pandas as pd 
import numpy as np 
pd.options.display.max_rows = 1000

## Child Data

# Read data of child and explore 
raw_df_child = pd.read_csv(DATA_PATH_CHILD)

%store raw_df_child

raw_df_child.sample(10)

- `Allele : is a variant of the sequence of nucleotides at a particular location, or locus, on a DNA molecule`
- visualize whats going on and what alleles are ?
- https://learn.genetics.utah.edu/content/genetics/whatallele

raw_df_child.info()

- `Name : is the generation of the individual with an Unique identifier `

- `Gender : is the sex of the individual (M or F)`

- `ParentM : is the mother of the individual`

- `ParentF : is the father of the individual`

- `EthA : is the ethnicity of the individual of Africa American`

- `EthE : is the ethnicity of the individual of Estonian`

- `EthK : is the ethnicity of the individual of Korean`

- `EthP : is the ethnicity of the individual of the Palestinian`

- `Allele1 : is a variant of the sequence of nucleotides at a particular location, or locus, on a DNA molecule`

- `Allele2 : is a variant of the sequence of nucleotides at a particular location, or locus, on a DNA molecule`

raw_df_child.isnull().sum()

allele1_information_child = raw_df_child['Allele1'][0]
len(allele1_information_child)

allele1_information_child

### **Explore more about child data**

# Display information of child of one Parent  (mother)
raw_df_child [raw_df_child['ParentM'] == 'A8184']

raw_df_child [raw_df_child['ParentF'] == 'A13904']

len(raw_df_child [raw_df_child['ParentM'] == 'A8184'])

- ` here we can see that data for parent has  A8184 has 8 records`

------

## Parent Data

raw_df_parent = pd.read_csv(DATA_PATH_PARENT)

%store raw_df_parent

raw_df_parent.sample(10)

----

- `ParentM and ParentF is Na for 1st generation individuals `

----

raw_df_parent.info()

raw_df_child.isnull().sum()

allele1_information_parent = raw_df_parent['Allele1'][0]

allele1_information_parent

raw_df_parent [raw_df_parent['Name'] == 'A13904']
