==> notebooks/01_preparation.ipynb <==
# <center> Data Preparation </center>
## <center> Step 1 <center>

------

# Constants
DATA_PATH_CHILD  = '../data/raw/G2000_data.csv'
DATA_PATH_PARENT = '../data/raw/EthA_data.csv'
EXPORT_PATH = '../data/processed/1_first_processed_merged_df.pkl'
FULL_SEQ = ['AA', 'AC', 'AG', 'AT', 'CA', 'CC', 'CG', 'CT', 'GA', 'GC', 'GG', 'GT', 'TA', 'TC', 'TG', 'TT']

import pandas as pd 
import numpy as np
import logging 
import pickle

raw_df_child = pd.read_csv(DATA_PATH_CHILD)

# Keeping copy of raw data
df_child = raw_df_child.copy()

raw_df_parent = pd.read_csv(DATA_PATH_PARENT)

df_parent = raw_df_parent.copy()

-----

<center> - Creating a new column in the DataFrame and assigning values based on concatenating the values from two existing columns <center>

<center>** full sequence of alleles **</center>

-----

# Childern full sequence of alleles
df_child['Child_full_DNA_Seq']= df_child['Allele1'] + df_child['Allele2']

df_parent['Parent_full_DNA_Seq']= df_parent['Allele1'] + df_parent['Allele2']

# make copy of G2000 ['ParentM','ParentF','Child_full_DNA_Seq']
df_child_data_copy = df_child[['ParentM','ParentF','Child_full_DNA_Seq']]

# Display information of (df_child_data_copy) G2000 ['ParentM','ParentF','Child_full_DNA_Seq']
df_child_data_copy.sample(10)

# make copy of EthA ['Name','Parent_full_DNA_Seq']
df_parent_data_copy = df_parent[['Name','Parent_full_DNA_Seq']]

df_parent_data_copy.sample(10)

-----

## **Merging Operation**

-----

**Merging :**

Merge the dataframes using the `Name` column as the key. The key column for merging is located in the `df_parent` dataframe.



**Keys in `df_child` DataFrame:**

- Left DataFrame Key (`left_on`): `Name`
- Right DataFrame Key (`right_on`): `Father or Mother`



------

# Merging Once with Father and Once with Mother
parent_child_merge_father = pd.merge(df_parent_data_copy , df_child_data_copy , left_on = 'Name' , right_on= 'ParentF')
parent_child_merge_mother = pd.merge(df_parent_data_copy , df_child_data_copy , left_on = 'Name' , right_on= 'ParentM')

# check each dataframe
parent_child_merge_father.sample(10)

-----

- ` here when we chech example row 1 we see that Name = 'A5643' and ParentF = 'A5643' and ParentM = 'A5036'`

- ` From this we can say that they are the same person`

- ` From data source they say that Name is unique identifier for a person`

- ` Finally, we conclude that Name is Father of the child ` 

------

parent_child_merge_mother.sample(10)

cleaned_data_first_part = parent_child_merge_father.drop(columns=['ParentM', 'ParentF'])
cleaned_data_first_part

first_merged_df= pd.concat([cleaned_data_first_part, parent_child_merge_mother])

-----

## Verify results 

-----

raw_df_parent [raw_df_parent['Name'] == 'A13904']

first_merged_df

-----

# Export Data 

-----

first_merged_df.to_pickle(EXPORT_PATH)
