==> notebooks/03_augmentation.ipynb <==
# <center> Data Augmentation </center>
-----
## <center> Step 2 <center>

-------

- we need to load more data for parents that are not in our dataset to overcome the limitation of our dataset also underfitting the model

# Constants
DATA_PATH_RAW = '../data/raw/EthE_data.csv'
DATA_PATH_PROCESSED = '../data/processed/2_second_processed_merged_df.pkl'
EXPORT_PATH = '../data/processed/final_processed_data.pkl'

import pandas as pd 
import numpy as np

e_raw_df = pd.read_csv(DATA_PATH_RAW)
e_raw_df.sample(4)

e_raw_df.info()

e_raw_df['Parent_full_DNA_Seq'] = e_raw_df['Allele1'] + e_raw_df['Allele2']

e_raw_df.sample(2)

e_raw_df_copy = e_raw_df[['Name', 'Parent_full_DNA_Seq']]

e_raw_df_copy['target'] = 0 

e_raw_df_copy.sample(4)

second_processed_merged_df = pd.read_pickle(DATA_PATH_PROCESSED)

final_processed_data = pd.concat([second_processed_merged_df , e_raw_df_copy])

final_processed_data

final_processed_data['target'].value_counts()

final_processed_data.to_pickle(EXPORT_PATH)
