==> notebooks/04_model_baseline.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer , CountVectorizer
#from sklearn.preprocessing import OneHotEncoder
from sklearn.svm import SVC
from sklearn.metrics import accuracy_score , precision_score , recall_score , f1_score , confusion_matrix
import xgboost as xgb
from xgboost import XGBClassifier
#from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.utils import resample
import pickle
import joblib

df_ = pd.read_pickle("../data/processed/final_processed_data.pkl")

df = df_.copy()

# db = df.head(1000) # for test 

# db.to_csv('../data/raw/Testdb2.csv', index=False)

df.drop(df.tail(100).index,inplace=True)
df

df['Parent_full_DNA_Seq'] = df['Parent_full_DNA_Seq'].astype('str')
df['Child_full_DNA_Seq'] = df['Child_full_DNA_Seq'].astype('str')

df_pre = df[['Parent_full_DNA_Seq','Child_full_DNA_Seq','target']]

df_pre

x = df_pre.drop('target',axis=1)
y = df_pre['target']
X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.3, random_state=42)

x.shape , y.shape

x

y

le = LabelEncoder()
y_train = le.fit_transform(y_train)

kmer_embeddings_c = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/train_vectorizer_np/kmer_embeddings_c1.npy')
kmer_embeddings_p = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/train_vectorizer_np/kmer_embeddings_p1.npy')

# # vectorizer = CountVectorizer(analyzer = 'char' ,ngram_range=(3,3) , min_df=1)
# vectorizer = joblib.load('../data/interim/kmer_model.sav')
# X_ch = vectorizer.fit_transform(X_train['Child_full_DNA_Seq'])
# X_p = vectorizer.fit_transform(X_train['Parent_full_DNA_Seq'])
# kmer_embeddings_c = X_ch.toarray()
# kmer_embeddings_p = X_p.toarray()

kmer_embeddings_p = np.pad(kmer_embeddings_p, ((0, 0), (0, 1)), 'constant')
for i in range(len(kmer_embeddings_c[0])):
    column_name_child = f'child_gene_k_{i}'
    column_name_parent = f'parent_gene_k_{i}'
    X_train[column_name_child] = [kmer_embeddings_c[j][i] for j in range(len(X_train))]
    X_train[column_name_parent] = [kmer_embeddings_p[j][i] for j in range(len(X_train))]

# Drop specific ['child_gene_k_64', 'parent_gene_k_64'] columns if present in train data 
columns_to_drop = ['child_gene_k_64', 'parent_gene_k_64']
X_train.drop(columns_to_drop, axis=1, errors='ignore', inplace=True)

X_train = X_train.drop(['Parent_full_DNA_Seq','Child_full_DNA_Seq'],axis=1)

kmer_embeddings_child_test = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/test_vectorizer_np/kmer_embeddings_child_test1.npy')
kmer_embeddings_parent_test = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/test_vectorizer_np/kmer_embeddings_parent_test1.npy')

# # Working with test data
# X_t_c= vectorizer.transform(X_test['Child_full_DNA_Seq'])
# # X_t_c vect
# X_t_p = vectorizer.transform(X_test['Parent_full_DNA_Seq'])
# # X_t_p >>>>>>>> vectorizer_0
# kmer_embeddings_child_test = X_t_c.toarray()
# kmer_embeddings_parent_test = X_t_p.toarray()

for i in range (0, len(kmer_embeddings_child_test[0])):
    X_test['child_gene_k_'+str(i)] = [kmer_embeddings_child_test[j][i] for j in range (0, len(X_test))]
    X_test['parent_gene_k_'+str(i)] = [kmer_embeddings_parent_test[j][i] for j in range (0, len(X_test))]

X_test.shape , X_train.shape

X_test.drop(['Parent_full_DNA_Seq','Child_full_DNA_Seq'],axis=1,inplace=True)

X_test

X_train

# X_train = X_train.drop(['child_gene_k_64','parent_gene_k_64'],axis=1)

model_xgb = xgb.XGBClassifier(
    learning_rate=0.3,
    n_estimators=15000,
    max_depth=15,
)


from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
# transform data
X_train= scaler.fit_transform(X_train)
X_test= scaler.transform(X_test)

#train_x = X_train.values
#test_x = X_test.values

# grid_search = GridSearchCV(estimator=ada_boost, param_grid=param_grid, scoring='accuracy', cv=5)
# grid_search.fit(X_train, y_train)

model_xgb.fit(X_train, y_train)
# 9 min 51 sec  test/ 1
# 30 min 22 sec  test/ 2
# 8 min 23 sec  test/ 3

y_pred_xgb = model_xgb.predict(X_test)

y_pred_xgb

accuracy = accuracy_score(y_test, y_pred_xgb)
precision = precision_score(y_test, y_pred_xgb)
recall = recall_score(y_test, y_pred_xgb)
f1 = f1_score(y_test, y_pred_xgb)
conf_matrix = confusion_matrix(y_test, y_pred_xgb)

print(f"Accuracy: {accuracy * 100:.2f}%")
print(f"Precision: {precision * 100:.2f}%")
print(f"Recall: {recall * 100:.2f}%")
print(f"F1 Score: {f1 * 100:.2f}%")
print("Confusion Matrix:")
print(conf_matrix)
print("\n")

# with open('model_baseline.pkl' , 'wb') as f:
#   pickle.dump(model_xgb, f)

## Validation of the model & Testing New Data

- i will extract the new data from the csv file
- i will use the model to predict the labels for the new data

df_tt = pd.read_csv('../data/raw/Testdb2.csv')

df_tt

df_tt.target.value_counts()

kmer_embeddings_child_test = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/test_vectorizer_np/kmer_embeddings_child_test1.npy')
kmer_embeddings_parent_test = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/test_vectorizer_np/kmer_embeddings_parent_test1.npy')

for i in range (0, len(kmer_embeddings_child_test[0])):
    df_tt['child_gene_k_'+str(i)] = [kmer_embeddings_child_test[j][i] for j in range (0, len(df_tt))]
    df_tt['parent_gene_k_'+str(i)] = [kmer_embeddings_parent_test[j][i] for j in range (0, len(df_tt))]

y_dd = df_tt['target']

df_tt.columns

df_tt.drop(['Name','Parent_full_DNA_Seq','Child_full_DNA_Seq','ParentM','ParentF', 'target'],axis=1,inplace=True)

df_tt

y_pred_ = model_xgb.predict(df_tt)

y_pred_

accuracy = accuracy_score(y_dd, y_pred_)
precision = precision_score(y_dd, y_pred_)
recall = recall_score(y_dd, y_pred_)
f1 = f1_score(y_dd, y_pred_)
conf_matrix = confusion_matrix(y_dd, y_pred_)

print(f"Accuracy: {accuracy * 100:.2f}%")
print(f"Precision: {precision * 100:.2f}%")
print(f"Recall: {recall * 100:.2f}%")
print(f"F1 Score: {f1 * 100:.2f}%")
print("Confusion Matrix:")
print(conf_matrix)
print("\n")

# Model Pipeline:

# from sklearn.pipeline import Pipeline
# from sklearn.compose import ColumnTransformer
# from sklearn.preprocessing import FunctionTransformer
# import pandas as pd
# import joblib

# # Function to create DataFrame
# def create_data_frame(child_seq, possible_parent_seq):
#     repeated_value = child_seq[0]
#     data = {
#         'possiable_child': [repeated_value] * len(possible_parent_seq),
#         'possiable_parent': possible_parent_seq
#     }
#     # Convert the dictionary to a pandas DataFrame
#     test_df = pd.DataFrame(data)
#     return test_df

# # Function to convert columns to categorical
# def to_categorical(df_):
#     df_ca = df_.copy()
#     df_ca['possiable_child'] = df_ca['possiable_child'].astype('category')
#     df_ca['possiable_parent'] = df_ca['possiable_parent'].astype('category')
#     return df_ca

# # Function to encode DataFrame
# def encode_df(df_):
#     df_c = df_.copy()
#     # Perform encoding using loaded_model or any other appropriate method
#     # Example: loaded_model.transform(df_c['possiable_child'])
#     # Example: loaded_model.transform(df_c['possiable_parent'])
#     # For demonstration purposes, let's assume loaded_model is defined elsewhere
#     # and applied as a TfidfVectorizer
#     loaded_model = joblib.load('../data/interim/kmer_model.sav')
#     X_t_c = loaded_model.fit_transform(df_c['possiable_child'])
#     X_t_p = loaded_model.fit_transform(df_c['possiable_parent'])
    
#     kmer_embeddings_child_test = X_t_c.toarray()
#     kmer_embeddings_parent_test = X_t_p.toarray()
#     for i in range (0, len(kmer_embeddings_child_test[0])):
#         df_c['child_gene_k_'+str(i)] = [kmer_embeddings_child_test[j][i] for j in range (0, len(df_c))]
#         df_c['parent_gene_k_'+str(i)] = [kmer_embeddings_parent_test[j][i] for j in range (0, len(df_c))]
        
#     return df_c.drop(['possiable_child' , 'possiable_parent'] , axis = 1)

# # Define the pipeline steps
# preprocessing_pipeline = ColumnTransformer([
#     ('encode_df', FunctionTransformer(encode_df), ['possiable_child', 'possiable_parent'])
# ])

# # Create the pipeline
# pipeline = Pipeline([
#     ('create_df', FunctionTransformer(create_data_frame, kw_args={'possible_parent_seq': possible_parent_seq})),
#     ('to_categorical', FunctionTransformer(to_categorical)),
#     ('preprocessing', preprocessing_pipeline)
# ])

# # Example usage:
# child_seq = 'ACTG'
# possible_parent_seq = ['ACTG', 'CGTA', 'GATC']
# processed_df = pipeline.transform(create_data_frame(child_seq, possible_parent_seq))
# print(processed_df.head())


# from sklearn.pipeline import Pipeline
# from sklearn.compose import ColumnTransformer
# from sklearn.preprocessing import FunctionTransformer
# import pandas as pd
# import joblib

# def create_data_frame(child_seq, possible_parent_seq):
#     repeated_value = child_seq[0]
#     data = {
#         'possiable_child': [repeated_value] * len(possible_parent_seq),
#         'possiable_parent': possible_parent_seq
#     }
#     # Convert the dictionary to a pandas DataFrame
#     test_df = pd.DataFrame(data)
#     return test_df


# # Function to convert columns to categorical
# def to_categorical(df_):
#     df_ca = df_.copy()
#     df_ca['possiable_child'] = df_ca['possiable_child'].astype('category')
#     df_ca['possiable_parent'] = df_ca['possiable_parent'].astype('category')
#     return df_ca

# # Function to encode DataFrame
# def encode_df(df_):
#     df_c = df_.copy()
#     loaded_model = joblib.load('../data/interim/kmer_model.sav')
#     X_t_c = loaded_model.fit_transform(df_c['possiable_child'])
#     X_t_p = loaded_model.fit_transform(df_c['possiable_parent'])
    
#     kmer_embeddings_child_test = X_t_c.toarray()
#     kmer_embeddings_parent_test = X_t_p.toarray()
#     for i in range (0, len(kmer_embeddings_child_test[0])):
#         df_c['child_gene_k_'+str(i)] = [kmer_embeddings_child_test[j][i] for j in range (0, len(df_c))]
#         df_c['parent_gene_k_'+str(i)] = [kmer_embeddings_parent_test[j][i] for j in range (0, len(df_c))]
        
#     return df_c.drop(['possiable_child' , 'possiable_parent'] , axis = 1)

# # Define the pipeline steps
# preprocessing_pipeline = ColumnTransformer([
#     ('encode_df', FunctionTransformer(encode_df), ['possiable_child', 'possiable_parent'])
# ])

# # Create the pipeline
# pipeline = Pipeline([
#     ('create_df', FunctionTransformer(create_data_frame, kw_args={'possible_parent_seq': possible_parent_seq})),
#     ('to_categorical', FunctionTransformer(to_categorical)),
#     ('preprocessing', preprocessing_pipeline)
# ])

# processed_df = pipeline.transform(create_data_frame(child_seq, possible_parent_seq))
# print(processed_df.head())


# from sklearn.pipeline import Pipeline
# from sklearn.compose import ColumnTransformer
# from sklearn.preprocessing import StandardScaler
# from sklearn.preprocessing import FunctionTransformer
# from sklearn.feature_extraction.text import CountVectorizer
# from xgboost import XGBClassifier
# import pandas as pd
# import numpy as np
# from sklearn.model_selection import train_test_split
# from sklearn.metrics import accuracy_score

# # Load data
# df_pre = pd.read_pickle("../data/processed/final_processed_data.pkl")
# x = df_pre.drop('target', axis=1)
# y = df_pre['target']
# X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.3, random_state=42)

# df_tt = pd.read_csv('../data/raw/Testdb2.csv')
# y_dd = df_tt['target']



# vectorizer = CountVectorizer(analyzer = 'char' ,ngram_range=(3,3) , min_df=1)
# X_ch = vectorizer.fit_transform(X_train['Child_full_DNA_Seq'])
# X_p = vectorizer.fit_transform(X_train['Parent_full_DNA_Seq'])
# kmer_embeddings_child = X_ch.toarray()
# kmer_embeddings_parent = X_p.toarray()

# # Define a function to extract kmer embeddings
# def extract_kmer_embeddings(X, kmer_embeddings_child, kmer_embeddings_parent):
#     X_copy = pd.DataFrame(X)  # Convert to DataFrame
#     for i in range(len(kmer_embeddings_child[0])):
#         X_copy['child_gene_k_' + str(i)] = [kmer_embeddings_child[j][i] for j in range(len(X_copy))]
#         X_copy['parent_gene_k_' + str(i)] = [kmer_embeddings_parent[j][i] for j in range(len(X_copy))]
#     return X_copy.drop(['Parent_full_DNA_Seq', 'Child_full_DNA_Seq'], axis=1)

# # Define the pipeline steps
# preprocessing_pipeline = Pipeline([
#     ('kmer_embeddings', FunctionTransformer(func=extract_kmer_embeddings, kw_args={'kmer_embeddings_child': kmer_embeddings_child, 'kmer_embeddings_parent': kmer_embeddings_parent})),
#     ('scale_numeric', ColumnTransformer([
#         ('scale', StandardScaler(), ['child_gene_k_' + str(i) for i in range(65)] + ['parent_gene_k_' + str(i) for i in range(65)])
#     ]))
# ])

# model_pipeline = Pipeline([
#     ('preprocessing', preprocessing_pipeline),
#     ('classifier', XGBClassifier(learning_rate=0.3, n_estimators=15000, max_depth=15))
# ])

# # Fit the pipeline
# model_pipeline.fit(X_train, y_train)

# # Predictions
# y_pred_test = model_pipeline.predict(X_test)
# y_pred_validation = model_pipeline.predict(df_tt)

# # Evaluate the model
# accuracy_test = accuracy_score(y_test, y_pred_test)
# accuracy_validation = accuracy_score(y_dd, y_pred_validation)

# print("Test Accuracy:", accuracy_test)
# print("Validation Accuracy:", accuracy_validation)


# from sklearn.pipeline import Pipeline
# from sklearn.compose import ColumnTransformer
# from sklearn.preprocessing import StandardScaler
# from sklearn.preprocessing import FunctionTransformer
# from sklearn.feature_extraction.text import CountVectorizer
# from xgboost import XGBClassifier
# import pandas as pd
# import numpy as np
# from sklearn.model_selection import train_test_split
# from sklearn.metrics import accuracy_score

# # Load data
# df_pre = pd.read_pickle("../data/processed/final_processed_data.pkl")
# x = df_pre.drop('target', axis=1)
# y = df_pre['target']
# X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.3, random_state=42)

# df_tt = pd.read_csv('../data/raw/Testdb2.csv')
# y_dd = df_tt['target']

# # Function to convert DNA sequences to strings
# def convert_to_string(df):
#     df_copy = df.copy()
#     df_copy['Child_full_DNA_Seq'] = df_copy['Child_full_DNA_Seq'].astype(str)
#     df_copy['Parent_full_DNA_Seq'] = df_copy['Parent_full_DNA_Seq'].astype(str)
#     return df_copy

# # Define a function to extract kmer embeddings
# def extract_kmer_embeddings(X, kmer_embeddings_child, kmer_embeddings_parent):
#     X_copy = pd.DataFrame(X)  # Convert to DataFrame
#     num_samples = len(X_copy)
#     num_features_child = len(kmer_embeddings_child[0])
#     num_features_parent = len(kmer_embeddings_parent[0])
    
#     # Check if the dimensions match
#     if num_samples != len(kmer_embeddings_child) or num_samples != len(kmer_embeddings_parent):
#         raise ValueError("Dimensions of input data and embeddings do not match.")
    
#     for i in range(num_features_child):
#         if i >= num_features_child or i >= num_features_parent:
#             raise IndexError("Index out of bounds.")
#         X_copy['child_gene_k_' + str(i)] = [kmer_embeddings_child[j][i] for j in range(num_samples)]
#         X_copy['parent_gene_k_' + str(i)] = [kmer_embeddings_parent[j][i] for j in range(num_samples)]
#     return X_copy.drop(['Parent_full_DNA_Seq', 'Child_full_DNA_Seq'], axis=1)


# # Define the pipeline steps
# preprocessing_pipeline = Pipeline([
#     ('convert_to_string', FunctionTransformer(func=convert_to_string)),
#     ('vectorize', ColumnTransformer([
#         ('count_vectorizer_child', CountVectorizer(analyzer='char', ngram_range=(3, 3), min_df=1), 'Child_full_DNA_Seq'),
#         ('count_vectorizer_parent', CountVectorizer(analyzer='char', ngram_range=(3, 3), min_df=1), 'Parent_full_DNA_Seq')
#     ])),
#     ('kmer_embeddings', FunctionTransformer(func=extract_kmer_embeddings, kw_args={'kmer_embeddings_child': kmer_embeddings_child, 'kmer_embeddings_parent': kmer_embeddings_parent})),
#     ('scale_numeric', ColumnTransformer([
#         ('scale', StandardScaler(), ['child_gene_k_' + str(i) for i in range(64)] + ['parent_gene_k_' + str(i) for i in range(64)])
#     ]))
# ])

# model_pipeline = Pipeline([
#     ('preprocessing', preprocessing_pipeline),
#     ('classifier', XGBClassifier(learning_rate=0.3, n_estimators=15000, max_depth=15))
# ])

# # Fit the pipeline
# model_pipeline.fit(X_train, y_train)

# # Predictions
# y_pred_test = model_pipeline.predict(X_test)
# y_pred_validation = model_pipeline.predict(df_tt)

# # Evaluate the model
# accuracy_test = accuracy_score(y_test, y_pred_test)
# accuracy_validation = accuracy_score(y_dd, y_pred_validation)

# print("Test Accuracy:", accuracy_test)
# print("Validation Accuracy:", accuracy_validation)


# import pandas as pd
# import numpy as np
# from sklearn.pipeline import Pipeline
# from sklearn.compose import ColumnTransformer
# from sklearn.preprocessing import StandardScaler
# from sklearn.preprocessing import FunctionTransformer
# from xgboost import XGBClassifier
# from sklearn.model_selection import train_test_split
# from sklearn.metrics import accuracy_score

# # Load data
# df_pre = pd.read_pickle("../data/processed/final_processed_data.pkl")
# x = df_pre.drop('target', axis=1)
# y = df_pre['target']
# X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.3, random_state=42)

# df_tt = pd.read_csv('../data/raw/Testdb2.csv')
# y_dd = df_tt['target']

# # Example embeddings arrays
# kmer_embeddings_child = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/test_vectorizer_np/kmer_embeddings_child_test.npy')
# kmer_embeddings_parent = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/test_vectorizer_np/kmer_embeddings_parent_test.npy')
# kmer_embeddings_child = np.pad(kmer_embeddings_child, ((0, 0), (0, 1)), 'constant')
# kmer_embeddings_parent = np.pad(kmer_embeddings_parent, ((0, 0), (0, 1)), 'constant')

# def extract_kmer_embeddings(X, kmer_embeddings_child, kmer_embeddings_parent):
#     X_copy = pd.DataFrame(X)  # Convert to DataFrame
#     num_samples = len(X_copy)
#     num_features_child = len(kmer_embeddings_child[0])
#     num_features_parent = len(kmer_embeddings_parent[0])
    
#     # Check if the dimensions match
#     if num_samples != len(kmer_embeddings_child) or num_samples != len(kmer_embeddings_parent):
#         raise ValueError("Dimensions of input data and embeddings do not match.")
    
#     for i in range(num_features_child):
#         if i >= num_features_child or i >= num_features_parent:
#             raise IndexError("Index out of bounds.")
#         X_copy['child_gene_k_' + str(i)] = [kmer_embeddings_child[j][i] for j in range(num_samples)]
#         X_copy['parent_gene_k_' + str(i)] = [kmer_embeddings_parent[j][i] for j in range(num_samples)]
    
#     return X_copy.drop(['Parent_full_DNA_Seq', 'Child_full_DNA_Seq'], axis=1), None, None

# # Define the pipeline steps
# preprocessing_pipeline = ColumnTransformer([
#     ('kmer_embeddings', FunctionTransformer(func=extract_kmer_embeddings, kw_args={'kmer_embeddings_child': kmer_embeddings_child, 'kmer_embeddings_parent': kmer_embeddings_parent}), ['Parent_full_DNA_Seq', 'Child_full_DNA_Seq']),
#     ('scale_numeric', StandardScaler(), ['child_gene_k_' + str(i) for i in range(64)] + ['parent_gene_k_' + str(i) for i in range(64)])
# ])

# model_pipeline = Pipeline([
#     ('preprocessing', preprocessing_pipeline),
#     ('classifier', XGBClassifier(learning_rate=0.3, n_estimators=15000, max_depth=15))
# ])

# # Fit the pipeline
# model_pipeline.fit(X_train, y_train)

# # Predictions
# y_pred_test = model_pipeline.predict(X_test)
# y_pred_validation = model_pipeline.predict(df_tt)

# # Evaluate the model
# accuracy_test = accuracy_score(y_test, y_pred_test)
# accuracy_validation = accuracy_score(y_dd, y_pred_validation)

# print("Test Accuracy:", accuracy_test)
# print("Validation Accuracy:", accuracy_validation)


# print(f"Accuracy: {accuracy * 100:.2f}%")
# print(f"Precision: {precision * 100:.2f}%")
# print(f"Recall: {recall * 100:.2f}%")
# print(f"F1 Score: {f1 * 100:.2f}%")
# print("Confusion Matrix:")
# print(conf_matrix)
# print("\n")

# ------------------
# 1- model_xgb.pkl
# Accuracy: 51.00%
# Precision: 50.78%
# Recall: 43.96%
# F1 Score: 47.12%
# Confusion Matrix:
# [[175 127]
#  [167 131]]
# ------------------
# 2- best_ada_boost_model.pkl
# Accuracy: 49.00%
# Precision: 49.23%
# Recall: 85.57%
# F1 Score: 62.50%
# Confusion Matrix:
# [[ 39 263]
#  [ 43 255]]
# ------------------
# 3- model_xgb_clf.pkl
# ValueError: y must have at least two dimensions for multi-output regression but has only one.
# ------------------
# 4- model_xgb_clf_hyperparameter.pkl
# ValueError: y must have at least two dimensions for multi-output regression but has only one.


# # saving Model
# with open('model_xgb.pkl' , 'wb') as f:
#   pickle.dump(model_xgb, f)
