==> notebooks/04_model_vector_Adaboost.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer , CountVectorizer
#from sklearn.preprocessing import OneHotEncoder
from sklearn.svm import SVC
from sklearn.metrics import accuracy_score , precision_score , recall_score , f1_score , confusion_matrix
import xgboost as xgb
from xgboost import XGBClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.utils import resample
import pickle
import joblib

df_ = pd.read_pickle("/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/processed/final_processed_data.pkl")

df = df_.copy()

df.drop(df.tail(1000).index,inplace=True)
df

data_0 = df[df['target'] == 0.0]
data_1 = df[df['target'] == 1.0]

from sklearn.utils import resample

data_0 = resample(data_0 ,replace=False ,n_samples=1000 ,random_state=42)
data_1 = resample(data_1 ,replace=False ,n_samples=1000 ,random_state=42)

df = pd.concat([data_0,data_1])
df.reset_index(inplace=True)
df.drop('index',axis=1,inplace=True)
df

df['Parent_full_DNA_Seq'] = df['Parent_full_DNA_Seq'].astype('str')
df['Child_full_DNA_Seq'] = df['Child_full_DNA_Seq'].astype('str')

df_pre = df[['Parent_full_DNA_Seq','Child_full_DNA_Seq','target']]

x = df_pre.drop('target',axis=1)
y = df_pre['target']
X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.3, random_state=42)

x.shape , y.shape

le = LabelEncoder()
y_train = le.fit_transform(y_train)

kmer_embeddings_c = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/train_vectorizer_np/kmer_embeddings_c.npy')
kmer_embeddings_p = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/train_vectorizer_np/kmer_embeddings_p.npy')

kmer_embeddings_p = np.pad(kmer_embeddings_p, ((0, 0), (0, 1)), 'constant')
for i in range (0 , len(kmer_embeddings_c[0])):
    X_train['child_gene_k_' + str(i)] = [kmer_embeddings_c[j][i] for j in range (0, len(X_train))]
    X_train['parent_gene_k_' + str(i)] = [kmer_embeddings_p[j][i] for j in range (0, len(X_train))]

X_train = X_train.drop(['Parent_full_DNA_Seq','Child_full_DNA_Seq'],axis=1)

param_grid = {
    'n_estimators': [15000],  # Adjust the values based on your requirements
    'learning_rate': [0.3],  # Adjust the values based on your requirements
    'base_estimator__max_depth': [15]  # If you're using DecisionTreeClassifier as base estimator
}

kmer_embeddings_child_test = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/test_vectorizer_np/kmer_embeddings_child_test.npy')
kmer_embeddings_parent_test = np.load('/Users/0xnrous/Developer/e2e_graduationPROJECT/e2e_graduationPROJECT_DS/data/interim/test_vectorizer_np/kmer_embeddings_parent_test.npy')

for i in range (0, len(kmer_embeddings_child_test[0])):
    X_test['child_gene_k_'+str(i)] = [kmer_embeddings_child_test[j][i] for j in range (0, len(X_test))]
    X_test['parent_gene_k_'+str(i)] = [kmer_embeddings_parent_test[j][i] for j in range (0, len(X_test))]

X_train = X_train.drop(['child_gene_k_64','parent_gene_k_64'],axis=1)

X_test.shape , X_train.shape

X_test.drop(['Parent_full_DNA_Seq','Child_full_DNA_Seq'],axis=1,inplace=True)

X_test

X_train

from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
# transform data
X_train= scaler.fit_transform(X_train)
X_test= scaler.transform(X_test)

# Create the AdaBoostClassifier
base_estimator = DecisionTreeClassifier() # You can change this base estimator based on your needs
ada_boost = AdaBoostClassifier(base_estimator=base_estimator)

#train_x = X_train.values
#test_x = X_test.values

grid_search = GridSearchCV(estimator=ada_boost, param_grid=param_grid, scoring='accuracy', cv=5)
grid_search.fit(X_train, y_train)

best_params = grid_search.best_params_
best_ada_boost_model = grid_search.best_estimator_

best_params

best_ada_boost_model

best_ada_boost_model.fit(X_train, y_train)

y_pred_ada_boost = best_ada_boost_model.predict(X_test)

y_pred_ada_boost

accuracy = accuracy_score(y_test, y_pred_ada_boost)
precision = precision_score(y_test, y_pred_ada_boost)
recall = recall_score(y_test, y_pred_ada_boost)
f1 = f1_score(y_test, y_pred_ada_boost)
conf_matrix = confusion_matrix(y_test, y_pred_ada_boost)

print(f"Accuracy: {accuracy * 100:.2f}%")
print(f"Precision: {precision * 100:.2f}%")
print(f"Recall: {recall * 100:.2f}%")
print(f"F1 Score: {f1 * 100:.2f}%")
print("Confusion Matrix:")
print(conf_matrix)
print("\n")

# Accuracy: 49.83%
# Precision: 49.70%
# Recall: 84.56%
# F1 Score: 62.61%
# Confusion Matrix:
# [[ 47 255]
#  [ 46 252]]

# saving Model
with open('best_ada_boost_model.pkl' , 'wb') as f:
  pickle.dump(best_ada_boost_model, f)
